db/init/data/requirements.txt
pandas
psycopg2-binary
pydantic
sqlalchemy